Refactor clock, weather, sidebar, bookmarks, shortcuts for better organization
