+++ components/css/bookmarks.css
@@ -1,0 +1,108 @@
+.shortcuts {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    max-width: 800px;
+    width: 100%;
+    font-size: 1.1rem;
+    text-align: left;
+    margin-left: 4rem;
+    max-height: calc(100vh - 18rem);
+    overflow-y: auto;
+    overflow-x: hidden;
+    padding-right: 1rem;
+    padding-bottom: 12rem;
+    -webkit-overflow-scrolling: touch;
+    overscroll-behavior: contain;
+    scrollbar-width: none; /* Hide scrollbar in Firefox */
+}
+
+.shortcuts::-webkit-scrollbar {
+    display: none; /* Hide scrollbar in WebKit browsers */
+}
+
+.bookmark-list {
+    padding-left: 14px;
+    padding-right: 24px;
+}
+
+.bookmark-list,
+.bookmark-children {
+    list-style: none;
+    margin: 0;
+    width: 100%;
+    max-width: 100%;
+    box-sizing: border-box;
+}
+
+.bookmark-folder-item,
+.bookmark-link-item {
+    width: 100%;
+    box-sizing: border-box;
+}
+
+.shortcut {
+    display: block;
+    margin: 0.2rem 0 0.4rem 0;
+    text-decoration: underline dotted;
+    text-underline-offset: 2px;
+    color: inherit;
+    cursor: pointer;
+    transition: color 0.2s ease, transform 0.2s ease;
+    word-break: break-word;
+    overflow-wrap: anywhere;
+    width: 100%;
+    max-width: 100%;
+    box-sizing: border-box;
+}
+
+.shortcut:hover {
+    color: #0077cc;
+    transform: translateX(2px);
+}
+
+.bookmark-folder {
+    font-weight: bold;
+    margin: 0.4rem 0 0.2rem 0;
+    cursor: pointer;
+    color: inherit;
+    user-select: none;
+    display: flex;
+    align-items: flex-start;
+    flex-wrap: wrap;
+    gap: 0.4rem;
+    width: 100%;
+    background: transparent;
+    border: none;
+    text-align: left;
+    padding: 0.2rem 0;
+    font: inherit;
+}
+
+.bookmark-folder .chevron,
+.bookmark-folder span {
+    white-space: normal;
+}
+
+.bookmark-folder span:last-child {
+    flex: 1;
+    min-width: 0;
+    word-break: break-word;
+    overflow-wrap: anywhere;
+}
+
+.bookmark-folder:hover {
+    color: #0077cc;
+}
+
+.bookmark-children {
+    padding-left: 1.2rem;
+    margin-top: 0.1rem;
+    overflow: visible;
+    max-height: none;
+    box-sizing: border-box;
+}
+
+.bookmark-children.collapsed {
+    display: none;
+}
