+++ components/css/clock.css
@@ -1,0 +1,4 @@
+.clock {
+    font-size: 5rem;
+    margin-bottom: 0.5rem;
+}
