+++ components/css/weather.css
@@ -1,0 +1,5 @@
+.weather {
+    font-size: 1.4rem;
+    margin-bottom: 2rem;
+    color: inherit;
+}
